fix: fixes some minor things
